ordenamos carpetas, falta arreglo de stats
