+++ src/main/java/com/github/cc3002/citricjuice/model/board/AbstractPanel.java
@@ -1,6 +1,6 @@
 package com.github.cc3002.citricjuice.model.board;
 
-import com.github.cc3002.citricjuice.model.Jugador;
+import com.github.cc3002.citricjuice.model.contenders.Jugador;
 import java.util.HashSet;
 import java.util.Set;
