+++ src/main/java/com/github/cc3002/citricjuice/model/board/HomePanel.java
@@ -1,6 +1,6 @@
 package com.github.cc3002.citricjuice.model.board;
 
-import com.github.cc3002.citricjuice.model.Jugador;
+import com.github.cc3002.citricjuice.model.contenders.Jugador;
 
 public class HomePanel extends AbstractPanel{
   public HomePanel(int idpanel) {
