--- src/main/java/com/github/cc3002/citricjuice/model/board/PanelType.java
@@ -1,12 +1,0 @@
-package com.github.cc3002.citricjuice.model.board;
-
-/**
- * Enum for the types of panels of the game.
- *
- * @author <a href="mailto:ignacio.slater@ug.uchile.cl">Ignacio Slater M.</a>.
- * @version 1.0.6-rc.1
- * @since 1.0
- */
-public enum PanelType {
-  BONUS, BOSS, DROP, ENCOUNTER, HOME, NEUTRAL
-}
