+++ src/main/java/com/github/cc3002/citricjuice/model/contenders/IContender.java
@@ -1,4 +1,4 @@
-package com.github.cc3002.citricjuice.model;
+package com.github.cc3002.citricjuice.model.contenders;
 
 public interface IContender {
   String getName();
