+++ src/main/java/com/github/cc3002/citricjuice/model/contenders/Wild.java
@@ -1,15 +1,14 @@
-package com.github.cc3002.citricjuice.model;
+package com.github.cc3002.citricjuice.model.contenders;
 
-public class Boss extends AbstractEnemy {
-
-  public Boss(String name, int hp, int atk, int def, int evd) {
+public class Wild extends AbstractEnemy{
+  public Wild(String name, int hp, int atk, int def, int evd) {
     super(name, hp, atk, def, evd);
   }
 
   @Override
   public void attack(IContender enemigo) {
     if (enemigo.getCurrentHP()!=0){
-      enemigo.attackedByBoss(this);
+      enemigo.attackedByWild(this);
     }
   }
   /**
@@ -18,34 +17,35 @@ public class Boss extends AbstractEnemy {
    * looking at
    */
 
+
   @Override
   public void loseByJugador(IContender jugador) {
     jugador.addStars(this.getStars());
-    jugador.addWins(3);
+    jugador.addWins(1);
     this.reduceStarsBy(this.getStars());
     //pierde todas las estrellas y se las pasa al jugador
-    //el jugador gana 3 victorias
+    //el jugador gana 1 victorias
   }
 
   @Override
   public void loseByBoss(IContender boss) {
     boss.addStars(this.getStars()/2);
-    boss.addWins(3);
+    boss.addWins(1);
     this.reduceStarsBy(this.getStars()/2);
     //debo perder la mitad de mis estrellas(y el boss ganarlas)
-    //el boss gana 3 victorias
+    //el boss gana 1 victorias
   }
 
   @Override
   public void loseByWild(IContender wild) {
     wild.addStars(this.getStars()/2);
-    wild.addWins(3);
+    wild.addWins(1);
     this.reduceStarsBy(this.getStars()/2);
     //debo perder la mitad de mis estrellas(y el wild ganarlas)
-    //el wild gana 3 victorias
+    //el wild gana 1 victorias
   }
 
-  public Boss copy() {
-    return new Boss(this.getName(), this.getMaxHP(), this.getAtk(), this.getDef(), this.getEvd());
+  public Wild copy() {
+    return new Wild(this.getName(), this.getMaxHP(), this.getAtk(), this.getDef(), this.getEvd());
   }
 }
